fold calculator logic into calculate
