--- src/logic/isNumber.js
@@ -1,3 +1,0 @@
-export default function isNumber(item) {
-  return /[0-9]+/.test(item);
-}
